fix(global): Fix listing issues - Force casting type int for auth->company_id

+++ app/Http/Controllers/Email/EmailIndexController.php
@@ -14,7 +14,7 @@ class EmailIndexController extends MainController
     public function index(Request $request)
     {
         $data['emails'] = (new Email())->getAllByCompanyId(
-            Auth::user()->company_id,
+            (int) Auth::user()->company_id,
             $request->search
         );
